+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-06T13:58:33.810Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-07T04:32:46.815Z")
     r"""The date and time the resource was last updated"""