+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6WX2GR79RX00K8W016G9E0H"
+        "01K6YF37ZK2G07P9T5TVPF7A9V"
     )
     r"""The id of the resource"""