+++ packages/orq-rc/RELEASES.md
@@ -3169,3 +3169,13 @@ Based on:
 - [python v3.14.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.11] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.11 - packages/orq-rc
+
+## 2025-10-07 04:34:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.635.1 (2.722.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.12] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.12 - packages/orq-rc